+++ pkg/translate/googletranslate/googletranslate.go
@@ -2,75 +2,77 @@ package googletranslate
 
 import (
 	"context"
-	"errors"
 	"fmt"
+	"io"
 
 	"cloud.google.com/go/translate"
-	"go.expect.digital/translate/pkg/model"
 	"golang.org/x/text/language"
+	"google.golang.org/api/option"
 )
 
-func (g *GoogleTranslate) validateTranslateReq(messages *model.Messages, targetLang language.Tag) error {
-	if messages == nil {
-		return errors.New("nil messages")
-	}
-
-	if len(messages.Messages) == 0 {
-		return errors.New("no messages")
-	}
-
-	if targetLang == language.Und {
-		return errors.New("target language undefined")
-	}
-
-	// Enforce that source language is supported.
-	if ok := g.supportedLangTags[messages.Language]; !ok {
-		return fmt.Errorf("source language %s not supported", messages.Language)
-	}
-
-	// Enforce that target language is supported.
-	if ok := g.supportedLangTags[targetLang]; !ok {
-		return fmt.Errorf("target language %s not supported", targetLang)
-	}
-
-	return nil
+type GoogleClient interface {
+	Translate(ctx context.Context, inputs []string, target language.Tag, opts *translate.Options) ([]translate.Translation, error) //nolint:lll
+	SupportedLanguages(ctx context.Context, target language.Tag) ([]translate.Language, error)
+	io.Closer
 }
 
-func (g *GoogleTranslate) Translate(
-	ctx context.Context,
-	messages *model.Messages,
-	targetLang language.Tag,
-) (*model.Messages, error) {
-	if err := g.validateTranslateReq(messages, targetLang); err != nil {
-		return nil, fmt.Errorf("google translate: validate translate request: %w", err)
+type GoogleTranslate struct {
+	client            GoogleClient
+	supportedLangTags map[language.Tag]bool
+}
+
+type TranslateOption func(*GoogleTranslate) error
+
+func WithClient(c GoogleClient) TranslateOption {
+	return func(g *GoogleTranslate) error {
+		g.client = c
+		return nil
+	}
+}
+
+func WithDefaultClient(ctx context.Context, apiKey string) TranslateOption {
+	return func(g *GoogleTranslate) error {
+		var err error
+
+		g.client, err = translate.NewClient(ctx, option.WithAPIKey(apiKey))
+		if err != nil {
+			return fmt.Errorf("with default client: new google translate client: %w", err)
+		}
+
+		return nil
+	}
+}
+
+func NewGoogleTranslate(ctx context.Context, opts ...TranslateOption) (*GoogleTranslate, func() error, error) {
+	googleTranslate := &GoogleTranslate{}
+
+	for _, opt := range opts {
+		if err := opt(googleTranslate); err != nil {
+			return nil, nil, fmt.Errorf("apply opt: %w", err)
+		}
 	}
 
-	targetTexts := make([]string, 0, len(messages.Messages))
-	for _, m := range messages.Messages {
-		targetTexts = append(targetTexts, m.ID)
-	}
-
-	// Set source language if defined, otherwise let Google Translate detect it.
-	opts := &translate.Options{}
-	if messages.Language != language.Und {
-		opts = &translate.Options{Source: messages.Language}
-	}
-
-	translations, err := g.client.Translate(ctx, targetTexts, targetLang, opts)
+	// Ping the Google Translate API to ensure that the client is working.
+	_, err := googleTranslate.client.Translate(ctx, []string{"Hello World!"}, language.Latvian, nil)
 	if err != nil {
-		return nil, fmt.Errorf("google translate client: translate: %w", err)
+		return nil, nil, fmt.Errorf("google translate client: ping google translate: %w", err)
 	}
 
-	translatedMessages := make([]model.Message, 0, len(translations))
-	for i, t := range translations {
-		translatedMessages = append(translatedMessages, model.Message{
-			ID:          messages.Messages[i].ID,
-			PluralID:    messages.Messages[i].PluralID,
-			Description: messages.Messages[i].Description,
-			Message:     t.Text,
-			Fuzzy:       true,
-		})
+	// Get the list of supported languages.
+	supported, err := googleTranslate.client.SupportedLanguages(ctx, language.English)
+	if err != nil {
+		return nil, nil, fmt.Errorf("google translate client: get supported languages: %w", err)
 	}
 
-	return &model.Messages{Language: messages.Language, Messages: translatedMessages}, nil
+	// Create a map of supported languages for quick lookup.
+	googleTranslate.supportedLangTags = make(map[language.Tag]bool, len(supported))
+	for _, lang := range supported {
+		googleTranslate.supportedLangTags[lang.Tag] = true
+	}
+
+	// Add the undefined language tag to the map of supported languages
+	// as Google Translate tries to detect the language then.
+	googleTranslate.supportedLangTags[language.Und] = true
+
+	return googleTranslate, googleTranslate.client.Close, nil
 }
